Fix: better numbering for last article of a text.
This time this fixes for instance a `**Art. 2** bis.` into `**Art.
2bis.`. This also fixes some monospace blocks.

One thing that may be a problem is that it removes leading lone dot `.`
characters. This is mostly a good thing but in some cases the dot is
used as a bullet point. E.g. ". soit:" in some texts.